chore: update uno sample app with latest
