+++ samples/CSharpUno/.vscode/settings.json
@@ -3,5 +3,8 @@
   "explorer.fileNesting.expand": false,
   "explorer.fileNesting.patterns": {
     "*.xaml": "$(capture).xaml.cs"
+  },
+  "files.associations": {
+    "global.json": "jsonc"
   }
 }
