+++ samples/CSharpUno/Directory.Build.props
@@ -12,12 +12,4 @@
     -->
     <NoWarn>$(NoWarn);NU1507;NETSDK1201;PRI257</NoWarn>
   </PropertyGroup>
-
-  <!-- See https://aka.platform.uno/using-uno-sdk#implicit-packages for more information regarding the Implicit Packages version properties. -->
-  <PropertyGroup>
-    <UnoExtensionsVersion>4.1.23</UnoExtensionsVersion>
-    <UnoToolkitVersion>6.0.24</UnoToolkitVersion>
-    <UnoThemesVersion>5.0.13</UnoThemesVersion>
-    <UnoCSharpMarkupVersion>5.2.14</UnoCSharpMarkupVersion>
-  </PropertyGroup>
 </Project>
