+++ samples/CSharpUno/Directory.Packages.props
@@ -3,12 +3,9 @@
     To update the version of Uno, you should instead update the Sdk version in the global.json file.
 
     See https://aka.platform.uno/using-uno-sdk for more information.
+    See https://aka.platform.uno/using-uno-sdk#implicit-packages for more information regarding the Implicit Packages.
   -->
   <ItemGroup>
-  </ItemGroup>
-  <ItemGroup>
-    <PackageVersion Include="System.Private.Uri" Version="4.3.2" />
-    <PackageVersion Include="System.Text.Json" Version="9.0.3" />
     <PackageVersion Include="Velopack" Version="0.0.1053" />
     <PackageVersion Include="Velopack.Build" Version="0.0.1053" />
   </ItemGroup>
