+++ samples/CSharpUno/UnoSample.sln
@@ -1,17 +1,18 @@
+﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.11.34929.205
+VisualStudioVersion = 17.13.35913.81 d17.13
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UnoSample", "UnoSample\UnoSample.csproj", "{F9ACED52-AE70-46F6-8C7C-2EF629084661}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UnoSample", "UnoSample\UnoSample.csproj", "{6BF41A07-4B32-447E-96AB-5437C3564AFD}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{29980CC9-DC41-4944-966D-F088C84D1C7B}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{02EA681E-C7D8-13C7-8484-4AC65E1B71E8}"
 	ProjectSection(SolutionItems) = preProject
+		.editorconfig = .editorconfig
 		.gitignore = .gitignore
 		Directory.Build.props = Directory.Build.props
 		Directory.Build.targets = Directory.Build.targets
 		Directory.Packages.props = Directory.Packages.props
 		global.json = global.json
-		README.md = README.md
 	EndProjectSection
 EndProject
 Global
@@ -20,15 +21,17 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F9ACED52-AE70-46F6-8C7C-2EF629084661}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F9ACED52-AE70-46F6-8C7C-2EF629084661}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F9ACED52-AE70-46F6-8C7C-2EF629084661}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F9ACED52-AE70-46F6-8C7C-2EF629084661}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6BF41A07-4B32-447E-96AB-5437C3564AFD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6BF41A07-4B32-447E-96AB-5437C3564AFD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6BF41A07-4B32-447E-96AB-5437C3564AFD}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
+		{6BF41A07-4B32-447E-96AB-5437C3564AFD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6BF41A07-4B32-447E-96AB-5437C3564AFD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6BF41A07-4B32-447E-96AB-5437C3564AFD}.Release|Any CPU.Deploy.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {2FA3FC94-E895-443E-AAA8-9A7F956F9E92}
+		SolutionGuid = {800ACAFE-BC3B-4646-9989-C59A522D41E4}
 	EndGlobalSection
 EndGlobal
