+++ samples/CSharpUno/UnoSample/GlobalUsings.cs
@@ -1,15 +1,3 @@
-﻿global using System.Collections.Immutable;
 global using CommunityToolkit.Mvvm.ComponentModel;
 global using CommunityToolkit.Mvvm.Input;
-global using Microsoft.Extensions.DependencyInjection;
-global using Microsoft.Extensions.Hosting;
-global using Microsoft.Extensions.Localization;
 global using Microsoft.Extensions.Logging;
-global using Microsoft.Extensions.Options;
-global using UnoSample.DataContracts;
-global using UnoSample.DataContracts.Serialization;
-global using UnoSample.Models;
-global using UnoSample.Presentation;
-global using UnoSample.Services.Caching;
-global using UnoSample.Services.Endpoints;
-global using ApplicationExecutionState = Windows.ApplicationModel.Activation.ApplicationExecutionState;
