--- samples/CSharpUno/UnoSample/Models/AppConfig.cs
@@ -1,6 +1,0 @@
-namespace UnoSample.Models;
-
-public record AppConfig
-{
-    public string? Environment { get; init; }
-}
