--- samples/CSharpUno/UnoSample/Models/Entity.cs
@@ -1,3 +1,0 @@
-namespace UnoSample.Models;
-
-public record Entity(string Name);
