+++ samples/CSharpUno/UnoSample/Package.appxmanifest
@@ -1,20 +1,12 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-
+<?xml version="1.0" encoding="utf-8"?>
 <Package
   xmlns="http://schemas.microsoft.com/appx/manifest/foundation/windows10"
   xmlns:uap="http://schemas.microsoft.com/appx/manifest/uap/windows10"
   xmlns:rescap="http://schemas.microsoft.com/appx/manifest/foundation/windows10/restrictedcapabilities"
   IgnorableNamespaces="uap rescap">
 
-  <Identity
-    Name="UnoSample"
-    Publisher="O=Velopack"
-    Version="1.0.0.0" />
-
-  <Properties>
-    <DisplayName>UnoSample</DisplayName>
-    <PublisherDisplayName>UnoSample</PublisherDisplayName>
-  </Properties>
+  <Identity />
+  <Properties />
 
   <Dependencies>
     <TargetDeviceFamily Name="Windows.Universal" MinVersion="10.0.17763.0" MaxVersionTested="10.0.19041.0" />
@@ -29,11 +21,7 @@
     <Application Id="App"
       Executable="$targetnametoken$.exe"
       EntryPoint="$targetentrypoint$">
-      <uap:VisualElements
-        DisplayName="UnoSample"
-        Description="UnoSample">
-        <uap:SplashScreen />
-      </uap:VisualElements>
+      <uap:VisualElements />
     </Application>
   </Applications>
