+++ samples/CSharpUno/UnoSample/Platforms/Desktop/Program.cs
@@ -3,21 +3,19 @@ using Uno.UI.Runtime.Skia;
 namespace UnoSample;
 public class Program
 {
-    [STAThread]
-    public static void Main(string[] args)
-    {
-#if (!useDependencyInjection && useLoggingFallback)
-        App.InitializeLogging();
+	[STAThread]
+	public static void Main(string[] args)
+	{
+		App.InitializeLogging();
 
-#endif
-        var host = SkiaHostBuilder.Create()
-            .App(() => new App())
-            .UseX11()
-            .UseLinuxFrameBuffer()
-            .UseMacOS()
-            .UseWindows()
-            .Build();
+		var host = SkiaHostBuilder.Create()
+			.App(() => new App())
+			.UseX11()
+			.UseLinuxFrameBuffer()
+			.UseMacOS()
+			.UseWindows()
+			.Build();
 
-        host.Run();
-    }
+		host.Run();
+	}
 }
