--- samples/CSharpUno/UnoSample/Platforms/MacCatalyst/Info.plist
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-	<dict>
-		<key>UIDeviceFamily</key>
-		<array>
-			<integer>2</integer>
-		</array>
-		<key>LSApplicationCategoryType</key>
-		<string>public.app-category.utilities</string>
-		<key>UISupportedInterfaceOrientations</key>
-		<array>
-			<string>UIInterfaceOrientationPortrait</string>
-			<string>UIInterfaceOrientationLandscapeLeft</string>
-			<string>UIInterfaceOrientationLandscapeRight</string>
-		</array>
-		<key>XSAppIconAssets</key>
-		<string>Assets.xcassets/icon.appiconset</string>
-
-		<!--
-		Adjust this to your application's encryption usage.
-		<key>ITSAppUsesNonExemptEncryption</key>
-		<false/>
-		-->
-	</dict>
-</plist>
