--- samples/CSharpUno/UnoSample/Platforms/MacCatalyst/Main.maccatalyst.cs
@@ -1,13 +1,0 @@
-using UIKit;
-
-namespace UnoSample.MacCatalyst;
-public class EntryPoint
-{
-    // This is the main entry point of the application.
-    public static void Main(string[] args)
-    {
-        // if you want to use a different Application Delegate class from "AppDelegate"
-        // you can specify it here.
-        UIApplication.Main(args, null, typeof(App));
-    }
-}
