--- samples/CSharpUno/UnoSample/Platforms/MacCatalyst/Media.xcassets/LaunchImages.launchimage/Contents.json
@@ -1,58 +1,0 @@
-{
-  "images": [
-    {
-      "orientation": "portrait",
-      "extent": "full-screen",
-      "minimum-system-version": "7.0",
-      "scale": "2x",
-      "size": "640x960",
-      "idiom": "iphone"
-    },
-    {
-      "orientation": "portrait",
-      "extent": "full-screen",
-      "minimum-system-version": "7.0",
-      "subtype": "retina4",
-      "scale": "2x",
-      "size": "640x1136",
-      "idiom": "iphone"
-    },
-    {
-      "orientation": "portrait",
-      "extent": "full-screen",
-      "minimum-system-version": "7.0",
-      "scale": "1x",
-      "size": "768x1024",
-      "idiom": "ipad"
-    },
-    {
-      "orientation": "landscape",
-      "extent": "full-screen",
-      "minimum-system-version": "7.0",
-      "scale": "1x",
-      "size": "1024x768",
-      "idiom": "ipad"
-    },
-    {
-      "orientation": "portrait",
-      "extent": "full-screen",
-      "minimum-system-version": "7.0",
-      "scale": "2x",
-      "size": "1536x2048",
-      "idiom": "ipad"
-    },
-    {
-      "orientation": "landscape",
-      "extent": "full-screen",
-      "minimum-system-version": "7.0",
-      "scale": "2x",
-      "size": "2048x1536",
-      "idiom": "ipad"
-    }
-  ],
-  "properties": {},
-  "info": {
-    "version": 1,
-    "author": ""
-  }
-}
