--- samples/CSharpUno/UnoSample/Platforms/Windows/Program.cs
@@ -1,23 +1,0 @@
-using Microsoft.UI.Dispatching;
-using Velopack;
-
-namespace UnoSample;
-//public static class Program
-//{
-//    // https://learn.microsoft.com/en-us/windows/apps/windows-app-sdk/migrate-to-windows-app-sdk/guides/applifecycle?WT.mc_id=DT-MVP-5003472#single-instancing-in-main-or-wwinmain
-//    [STAThread]
-//    public static void Main(string[] args)
-//    {
-
-
-//        WinRT.ComWrappersSupport.InitializeComWrappers();
-
-//        Application.Start((p) =>
-//        {
-//            var context = new DispatcherQueueSynchronizationContext(
-//                DispatcherQueue.GetForCurrentThread());
-//            SynchronizationContext.SetSynchronizationContext(context);
-//            new App();
-//        });
-//    }
-//}
