--- samples/CSharpUno/UnoSample/Presentation/MainPage.xaml
@@ -1,39 +1,0 @@
-﻿<Page
-    x:Class="UnoSample.Presentation.MainPage"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:UnoSample.Presentation"
-    xmlns:uen="using:Uno.Extensions.Navigation.UI"
-    xmlns:um="using:Uno.Material"
-    xmlns:utu="using:Uno.Toolkit.UI"
-    Background="{ThemeResource BackgroundBrush}"
-    NavigationCacheMode="Required">
-
-    <Grid utu:SafeArea.Insets="VisibleBounds">
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto" />
-            <RowDefinition />
-        </Grid.RowDefinitions>
-        <utu:NavigationBar Content="{Binding Title}" />
-
-        <StackPanel
-            Grid.Row="1"
-            HorizontalAlignment="Center"
-            VerticalAlignment="Center"
-            Spacing="16">
-            <TextBlock Text="The app is not installed with Velopack" Visibility="{x:Bind NotInstalled}" />
-            <TextBlock Text="{x:Bind ViewModel.CurrentVersion}" />
-
-            <Button
-                HorizontalAlignment="Center"
-                Command="{Binding CheckForUpdatesCommand}"
-                Content="Check For Updates" />
-            <Button
-                HorizontalAlignment="Center"
-                Command="{Binding DownloadUpdatesCommand}"
-                Content="Download Updates"
-                Visibility="{x:Bind ViewModel.HasUpdate, Mode=OneWay}" />
-            <TextBlock Text="{Binding Status}" />
-        </StackPanel>
-    </Grid>
-</Page>
