--- samples/CSharpUno/UnoSample/Presentation/Shell.xaml.cs
@@ -1,10 +1,0 @@
-﻿namespace UnoSample.Presentation;
-
-public sealed partial class Shell : UserControl, IContentControlProvider
-{
-    public Shell()
-    {
-        this.InitializeComponent();
-    }
-    public ContentControl ContentControl => Splash;
-}
