+++ samples/CSharpUno/UnoSample/Properties/PublishProfiles/win-arm64.pubxml
@@ -1,22 +1,18 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <!--
 https://go.microsoft.com/fwlink/?LinkID=208121.
 -->
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
 	<PropertyGroup>
 		<PublishProtocol>FileSystem</PublishProtocol>
-		<Platform>x86</Platform>
-		<RuntimeIdentifier>win-x86</RuntimeIdentifier>
+		<Platform>arm64</Platform>
+		<RuntimeIdentifier>win-arm64</RuntimeIdentifier>
 		<PublishDir>bin\$(Configuration)\$(TargetFramework)\$(RuntimeIdentifier)\publish\</PublishDir>
 		<SelfContained>true</SelfContained>
 		<PublishSingleFile>False</PublishSingleFile>
 		<PublishReadyToRun Condition="'$(Configuration)' == 'Debug'">False</PublishReadyToRun>
 		<PublishReadyToRun Condition="'$(Configuration)' != 'Debug'">True</PublishReadyToRun>
-		<!-- Note: Trimming disabled by default as there may still be an issues with PublishTrimmed support: https://github.com/microsoft/CsWinRT/issues/373 -->
-		<!-- 
+		<PublishTrimmed Condition="'$(Configuration)' == 'Debug'">False</PublishTrimmed>
 		<PublishTrimmed Condition="'$(Configuration)' != 'Debug'">True</PublishTrimmed>
-		<TrimMode>partial</TrimMode>
-		<SuppressTrimAnalysisWarnings>true</SuppressTrimAnalysisWarnings> 
-		-->
 	</PropertyGroup>
 </Project>
