+++ samples/CSharpUno/UnoSample/Properties/PublishProfiles/win-x64.pubxml
@@ -12,11 +12,7 @@ https://go.microsoft.com/fwlink/?LinkID=208121.
 		<PublishSingleFile>False</PublishSingleFile>
 		<PublishReadyToRun Condition="'$(Configuration)' == 'Debug'">False</PublishReadyToRun>
 		<PublishReadyToRun Condition="'$(Configuration)' != 'Debug'">True</PublishReadyToRun>
-		<!-- Note: Trimming disabled by default as there may still be an issues with PublishTrimmed support: https://github.com/microsoft/CsWinRT/issues/373 -->
-		<!-- 
+		<PublishTrimmed Condition="'$(Configuration)' == 'Debug'">False</PublishTrimmed>
 		<PublishTrimmed Condition="'$(Configuration)' != 'Debug'">True</PublishTrimmed>
-		<TrimMode>partial</TrimMode>
-		<SuppressTrimAnalysisWarnings>true</SuppressTrimAnalysisWarnings> 
-		-->
 	</PropertyGroup>
 </Project>
