--- samples/CSharpUno/UnoSample/Services/Caching/IWeatherCache.cs
@@ -1,6 +1,0 @@
-namespace UnoSample.Services.Caching;
-
-public interface IWeatherCache
-{
-    ValueTask<IImmutableList<WeatherForecast>> GetForecast(CancellationToken token);
-}
