--- samples/CSharpUno/UnoSample/Services/Endpoints/IApiClient.cs
@@ -1,9 +1,0 @@
-using Refit;
-
-namespace UnoSample.Services.Endpoints;
-[Headers("Content-Type: application/json")]
-public interface IApiClient
-{
-    [Get("/api/weatherforecast")]
-    Task<ApiResponse<IImmutableList<WeatherForecast>>> GetWeather(CancellationToken cancellationToken = default);
-}
