--- samples/CSharpUno/UnoSample/Styles/ColorPaletteOverride.json
@@ -1,76 +1,0 @@
-{
-    "description": "Default Uno Material colors",
-    "seed": "#5946d2",
-    "coreColors": {
-        "primary": "#5946d2",
-        "secondary": "#6b4ea2"
-    },
-    "extendedColors": [],
-    "schemes": {
-        "light": {
-            "primary": "#5946d2",
-            "surfaceTint": "#5946d2",
-            "onPrimary": "#ffffff",
-            "primaryContainer": "#e5deff",
-            "onPrimaryContainer": "#170065",
-            "secondary": "#6b4ea2",
-            "onSecondary": "#ffffff",
-            "secondaryContainer": "#ebddff",
-            "onSecondaryContainer": "#220555",
-            "tertiary": "#0061a4",
-            "onTertiary": "#ffffff",
-            "tertiaryContainer": "#cfe4ff",
-            "onTertiaryContainer": "#001d36",
-            "error": "#b3261e",
-            "onError": "#ffffff",
-            "errorContainer": "#f9dedc",
-            "onErrorContainer": "#410e0b",
-            "background": "#fcfbff",
-            "onBackground": "#1c1b1f",
-            "surface": "#ffffff",
-            "onSurface": "#1c1b1f",
-            "surfaceVariant": "#f2eff5",
-            "onSurfaceVariant": "#8b8494",
-            "outline": "#79747e",
-            "outlineVariant": "#c9c5d0",
-            "shadow": "#000000",
-            "scrim": "#000000",
-            "inverseSurface": "#e6e1e5",
-            "inverseOnSurface": "#1c1b1f",
-            "inversePrimary": "#2a009f"
-        },
-        "dark": {
-            "primary": "#c7bfff",
-            "surfaceTint": "#c7bfff",
-            "onPrimary": "#2a009f",
-            "primaryContainer": "#4129ba",
-            "onPrimaryContainer": "#e4dfff",
-            "secondary": "#cdc2dc",
-            "onSecondary": "#332d41",
-            "secondaryContainer": "#433c52",
-            "onSecondaryContainer": "#eddfff",
-            "tertiary": "#9fcaff",
-            "onTertiary": "#003258",
-            "tertiaryContainer": "#00497d",
-            "onTertiaryContainer": "#d1e4ff",
-            "error": "#ffb4ab",
-            "onError": "#690005",
-            "errorContainer": "#93000a",
-            "onErrorContainer": "#ffdad6",
-            "background": "#1c1b1f",
-            "onBackground": "#e5e1e6",
-            "surface": "#302d37",
-            "onSurface": "#e6e1e5",
-            "surfaceVariant": "#47464f",
-            "onSurfaceVariant": "#c9c5d0",
-            "outline": "#928f99",
-            "outlineVariant": "#57545d",
-            "shadow": "#000000",
-            "scrim": "#000000",
-            "inverseSurface": "#e6e1e5",
-            "inverseOnSurface": "#1c1b1f",
-            "inversePrimary": "#2a009f"
-        }
-    },
-    "palettes": {}
-}
