--- samples/CSharpUno/UnoSample/appsettings.development.json
@@ -1,9 +1,0 @@
-﻿{
-  "AppConfig": {
-    "Environment": "Development"
-  },
-  "ApiClient": {
-    "Url": "https://localhost:5002",
-    "UseNativeHandler": true
-  }
-}
