--- samples/CSharpUno/UnoSample/appsettings.json
@@ -1,16 +1,0 @@
-{
-  "AppConfig": {
-    "Environment": "Production"
-  },
-  "ApiClient": {
-    "UseNativeHandler": true
-  },
-  "LocalizationConfiguration": {
-    "Cultures": [
-      "es",
-      "fr",
-      "pt-BR",
-      "en"
-    ]
-  }
-}
