+++ samples/CSharpUno/global.json
@@ -2,5 +2,8 @@
   // To update the version of Uno please update the version of the Uno.Sdk here. See https://aka.platform.uno/upgrade-uno-packages for more information.
   "msbuild-sdks": {
     "Uno.Sdk": "5.6.51"
+  },
+  "sdk":{
+    "allowPrerelease": false
   }
 }
